Generic device: Add vxWorks support
This requires some true wickedness.  Alas, the standard vxWorks compiler
is too old to provide C99 support, and the newer compiler that we already
use for fast feedback doesn't support C++ (due to binary vxWorks library
incompatibility), so we need to mix compilers.  Amazingly, this seems to
work!

Further annoyances are down to missing functionality in the vxWorks
version of libc, for instance C99 support is incomplete, and a couple of
functions have incompatible definitions, and some weaknesses in the
compiler which need workarounds.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/support/epics_device@126955 e099a375-04f9-0310-9d5f-a741eaff62e1