Add a mechanism for getting at the currently processed record
During record processing it is sometime useful to be able to call the
set_record_severity and set_record_timestamp functions, but without this
commit there is no straightforward way to discover the current record
without explicitly placing it in the record context.

This commit adds the use of thread local storage.  Fortunately, on modern
processors this is reasonably cheap.