Make process argument to _write_out_record work for waveforms
Turns out that the disable_write flag needs to be shared between between
both out *and* waveform records, and indeed we were incorrectly writing
this flag even when not using the OUT record type.

We only got away with this because the disable_write flag wasn't
overlapping with any important fields in the other union options.